pyproject.toml
[tool.poetry]
name = "atomic-habbits-drf"
version = "0.1.0"
description = ""
authors = ["korchaginks <kkonstantinserg@gmail.com>"]
readme = "README.md"
packages = [{include = "atomic_habbits_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
coverage = "^7.3.2"
pillow = "^10.1.0"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
requests = "^2.31.0"
eventlet = "^0.33.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
